% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{accuracy_profile}
\alias{accuracy_profile}
\title{Calculate accuracy, sensitivity, and the like.}
\usage{
accuracy_profile(logic_table)
}
\arguments{
\item{logic_table}{A list with names TP, FP, TN, FN.}
}
\value{
A list with `(acc, sens, spec, ppv, npv)`
}
\description{
This takes a truth table and returns accuracy,
sensitivity, specificity, positive predictive
validity, and negative predictive validity.
The truth table is a count of true positive, false positive,
true negative, and false negative.
}
\seealso{
\link{truth_table}
}
