% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{assign_to_grid}
\alias{assign_to_grid}
\title{Given spatial points of population, assign them to shapes.}
\usage{
assign_to_grid(grid, points)
}
\arguments{
\item{grid}{A regular grid.}

\item{points}{An sf points dataset.}
}
\value{
An sf grid with points assigned or
    NA where there are no points.
}
\description{
This transforms the grid into the same space as the points
dataset and then assigns the sum of points to the grid.
}
