% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{bimep_named_vector}
\alias{bimep_named_vector}
\title{Associate BIMEP index with its map ID.}
\usage{
bimep_named_vector(grid_100m)
}
\arguments{
\item{grid_100m}{The grid with `mapaID` defined.}
}
\value{
A vector of indexes and associated names.
}
\description{
Associate BIMEP index with its map ID.
}
