% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{bimep_on_grid}
\alias{bimep_on_grid}
\title{Aligns long-lat per-house data to a given raster grid.}
\usage{
bimep_on_grid(grid, bioko_sf, local_directory = "inst/extdata")
}
\arguments{
\item{grid}{A raster::raster to which to align}

\item{bioko_sf}{A shapefile that defines where population is zero.}

\item{local_directory}{Location of the directory where the file is found.}
}
\value{
A raster::raster with the new data on longlat, na in ocean,
  zero or a pop value on land.
}
\description{
Aligns long-lat per-house data to a given raster grid.
}
