% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{bimep_population_as_points}
\alias{bimep_population_as_points}
\title{Turns each population entry into a lat-long coordinate.}
\usage{
bimep_population_as_points(local_directory = "inst/extdata")
}
\arguments{
\item{local_directory}{Where external data is stored.}
}
\value{
An sf points file with population as features.
}
\description{
This reads the `popm.csv` of population at 100m grid points.
It produces a vector file with population as a feature.
}
