% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{bimep_population_as_raster}
\alias{bimep_population_as_raster}
\title{Read BIMEP population as a raster.}
\usage{
bimep_population_as_raster(grid_info, lat_long_proj, point_population)
}
\description{
This method shows that we really don't have the same raster as
the original BIMEP cells. This grid and the BIMEP grid aren't aligned.
}
\details{
We don't have the original raster, unfortunately, so we try
to rebuild it here. The data is organized into sections and then
areas within the section.
}
