% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter.R
\name{clean_grid}
\alias{clean_grid}
\title{Given an input grid, crop it to the admin unit and give us a few pixels of border.}
\usage{
clean_grid(grid, admin_sf)
}
\arguments{
\item{grid}{A raster to which to align.}

\item{admin_sf}{An sf shapefile that is the admin for cropping.}
}
\value{
A raster that is ready to use as a template.
}
\description{
Given an input grid, crop it to the admin unit and give us a few pixels of border.
}
