% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{data_configuration}
\alias{data_configuration}
\title{Reads configuration file on where to download data.}
\usage{
data_configuration()
}
\value{
A list of configuration parameters.
}
\description{
It stores that configuration in `data_config`
at the global level.
}
\details{
If you want to make the config file, then create
a file called `$HOME/.config/MASH/data.ini` and put
the following in it:
}
