% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{density_estimated}
\alias{density_estimated}
\title{Create a map with density estimation}
\usage{
density_estimated(population, bioko_sf, bandwidth, urban_cutoff = 10)
}
\arguments{
\item{bioko_sf}{Shapefile for bioko borders}

\item{bandwidth}{A bandwidth to use for the kernel density estimator
You have to choose this number. Take a look at `power_bandwidth`.}

\item{population_array}{a Raster of population}

\item{cutoff}{The minimum that determines}
}
\value{
A fraction that are above that level
}
\description{
Calculating urban fraction by pixel is strange when
you have 100 m pixels because urban fraciton is 1000
people in a square kilometer, which is 10 people in
a square 100 m, so it's more than one large household.
}
\details{
Therefore, we use a kernel density estimator for these
grid resolutions.
}
