% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{density_from_disc}
\alias{density_from_disc}
\title{Find the average density within a 1 square km disc around each pixel}
\usage{
density_from_disc(projected_raster)
}
\arguments{
\item{projected_raster}{a raster::raster that is projected to
have meters as units.}
}
\value{
A raster::raster that is similarly projected but now smoothed.
}
\description{
Given a raster that is projected to a grid measured in meters,
take each pixel and average its values with all pixels within
a disc that is 1 square kilometer in area.
}
