% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_hrsl_points}
\alias{download_hrsl_points}
\title{Download HRSL from the repository.}
\usage{
download_hrsl_points(local_directory = "inst/extdata")
}
\arguments{
\item{local}{directory Where to put that file on the local machine.}
}
\description{
HRSL is a geotiff in WGS 84 for all of Equatorial Guinea (GNQ).
}
