% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{filenames_to_description}
\alias{filenames_to_description}
\title{Given filenames, decode them into traits of the file.}
\usage{
filenames_to_description(filenames)
}
\arguments{
\item{filenames}{A character vector of filenames}
}
\value{
A dataframe with parameters of the files.
}
\description{
Given filenames, decode them into traits of the file.
}
