% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{geolocated_on_grid}
\alias{geolocated_on_grid}
\title{Aligns long-lat per-house data to a given raster grid.}
\usage{
geolocated_on_grid(grid, gelocated, admin_sf)
}
\arguments{
\item{grid}{A raster::raster to which to align}

\item{admin_sf}{A shapefile that defines where population is zero.}

\item{geolocated}{GPS data that's exact}
}
\value{
A raster::raster with the new data on longlat, na in ocean,
  zero or a pop value on land.
}
\description{
Aligns long-lat per-house data to a given raster grid.
}
