% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{hrsl_points}
\alias{hrsl_points}
\title{Given an HRSL raster, convert it into points in space.}
\usage{
hrsl_points(hrsl_raster)
}
\arguments{
\item{hsrl_raster}{This is a raster read from the HRSL geotiff.}
}
\value{
An sf points dataset with populations on the points.
}
\description{
Given an HRSL raster, convert it into points in space.
}
