% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{pareto_fraction}
\alias{pareto_fraction}
\title{Compute pareto fraction of a vector.}
\usage{
pareto_fraction(density)
}
\arguments{
\item{density}{This is a sortable, summable vector.}
}
\description{
This isn't the statistical estimator for a Pareto density
function's parameters. It calculates a Pareto principle,
the 80-20 rule sort of thing. The Pareto density function
assumes a minimum value, and this applies to any distribution.
}
