% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{population_density_estimator}
\alias{population_density_estimator}
\title{Estimate density of population using density estimator on points.}
\usage{
population_density_estimator(projected_raster)
}
\arguments{
\item{projected_raster}{A raster in projected coordinates.}
}
\value{
A list with a density raster and `m2_per_pixel`, which
  is square meters per pixel.
}
\description{
This takes a grid of population in latitude and longitude
and turns it into a point per person in order to estimate
the density of people.
}
