% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{read_bimep_point_data}
\alias{read_bimep_point_data}
\title{Read BIMEP point data}
\usage{
read_bimep_point_data(local_directory = "inst/extdata")
}
\arguments{
\item{local_directory}{where external data is stored.}
}
\description{
BIMEP point data is longitude-latitude data with people
in each house. This is only available within the collaboration.
}
