% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_bioko_grids}
\alias{read_bioko_grids}
\title{Read the Bioko grid information as a shapefile.}
\usage{
read_bioko_grids(local_directory = "inst/extdata")
}
\arguments{
\item{local}{directory Where to put that file on the local machine.}
}
\value{
a list with the `fine` and `coarse` shapefiles.
}
\description{
Read the Bioko grid information as a shapefile.
}
