% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_hrsl}
\alias{read_hrsl}
\title{Read the HRSL as a raster file.}
\usage{
read_hrsl(local_directory = "inst/extdata")
}
\arguments{
\item{local}{directory Where to put that file on the local machine.}
}
\value{
a `raster::raster` file.
}
\description{
Read the HRSL as a raster file.
}
