% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_landscan}
\alias{read_landscan}
\title{Load LandScan raster file.}
\usage{
read_landscan(local_directory = "inst/extdata")
}
\arguments{
\item{local}{directory Where to put that file on the local machine.}
}
\value{
a `raster::raster` file.
}
\description{
Load LandScan raster file.
}
