% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_worldpop}
\alias{read_worldpop}
\title{Load WorldPop raster file.}
\usage{
read_worldpop(local_directory = "inst/extdata")
}
\arguments{
\item{local}{directory Where to put that file on the local machine.}
}
\value{
a `raster::raster` file.
}
\description{
Load WorldPop raster file.
}
