% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{save_image}
\alias{save_image}
\title{Save an image to the images directory as both pdf and png.}
\usage{
save_image(name)
}
\arguments{
\item{name}{The base name to use for the file.}
}
\value{
The filenames to which it saved.
}
\description{
This saves the last ggplot image to a directory
as both pdf and png.
}
