% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot by base::plot to a file.}
\usage{
save_plot(plot_function, name)
}
\arguments{
\item{plot_function}{The function to run in order to plot}

\item{name}{The base name to use for the file.}
}
\value{
The filenames to which it saved.
}
\description{
This saves the plot to a directory
as both pdf and png.
}
