% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{square_meters_per_pixel.im}
\alias{square_meters_per_pixel.im}
\title{Calculate square meters per pixel for a spatstat im}
\usage{
square_meters_per_pixel.im(spatstat_im)
}
\arguments{
\item{spatstat_im}{The raster image in spatstat im format.}
}
\value{
A single value for square meters per pixel
}
\description{
Calculate square meters per pixel for a spatstat im
}
