% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{square_meters_per_pixel.raster}
\alias{square_meters_per_pixel.raster}
\title{Calculate square meters per pixel for a raster::raster}
\usage{
square_meters_per_pixel.raster(pop_raster)
}
\arguments{
\item{spatstat_im}{The raster image in raster::raster format.}
}
\value{
A single value for square meters per pixel
}
\description{
Calculate square meters per pixel for a raster::raster
}
