% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter.R
\name{summarize}
\alias{summarize}
\title{Compute summary statistics for geolocated data aligned to a grid.}
\usage{
summarize(geolocated, admin_sf, grid, shift, urban_cutoff)
}
\arguments{
\item{geolocated}{the GPS gold standard data.}

\item{admin_sf}{The boundary of the administrative unit as an sf shapefile.}

\item{grid}{A raster::raster to which to align the geolocated data.}

\item{shift}{A length-2 vector with the fraction of a pixel to shift the grid in x and y.}

\item{urban_cutoff}{A density above which summary statistics will call a place urban.}
}
\value{
A list of summary statistics.
}
\description{
Compute summary statistics for geolocated data aligned to a grid.
}
