% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter.R
\name{summary_statistics_multiple_grids}
\alias{summary_statistics_multiple_grids}
\title{Script to compute summary statistics with uncertainty from alignment of grid.}
\usage{
summary_statistics_multiple_grids(grid, shift_cnt, output)
}
\arguments{
\item{grid}{This says which grid to use, among hrsl, landscan, or worldpop.}

\item{shifts}{How many times to shift the grid in x and y. The number of loops internally
will be shifts * shifts.}

\item{filename}{A file to which to save results.}
}
\description{
Given gold standard data, which has lat-long GPS coordinates for each household,
align it to a grid and compute summary statistics for it, such as percent empty, percent urban,
and maximum. Compute an uncertainty in these values by repeatedly shifting
the grid so that we know how much slight translations of the grid can
affect results.
}
