% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{truth_table}
\alias{truth_table}
\title{Constructs a count of how many entries are in which truth table quadrant.}
\usage{
truth_table(condition, expected, observed)
}
\arguments{
\item{condition}{A function which, when applied to inputs, gives true and false.}

\item{expected}{The gold standard for what's true and false.}

\item{observed}{What is observed as true and false.}
}
\description{
Constructs a count of how many entries are in which truth table quadrant.
}
\seealso{
\link{accuracy_profile}
}
