% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{urban_fraction}
\alias{urban_fraction}
\title{Urban fraction from a density raster}
\usage{
urban_fraction(density_raster, urban_per_kilometer_sq)
}
\arguments{
\item{urban_per_kilometer_sq}{A number of people to define urban (1000, 1500).}

\item{pop_raster}{A raster::raster in projection}
}
\value{
A vector with c(numerator, denominator) of pixel counts.
}
\description{
Urban fraction from a density raster
}
