% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{urban_fraction_by_point_density}
\alias{urban_fraction_by_point_density}
\title{Urban fraction calculated by converting people to a point pattern.}
\usage{
urban_fraction_by_point_density(pop_raster, projection, urban_per_kilometer_sq)
}
\arguments{
\item{pop_raster}{A raster on lat-long.}

\item{projection}{A proj.4 string for projection.}

\item{urban_per_kilometer_sq}{A number of people to define urban (1000, 1500).}
}
\value{
A single density fraction.
}
\description{
Urban fraction calculated by converting people to a point pattern.
}
