% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{urban_fraction_density}
\alias{urban_fraction_density}
\title{Urban fraction from a density raster}
\usage{
urban_fraction_density(density_raster, urban_per_kilometer_sq)
}
\arguments{
\item{urban_per_kilometer_sq}{A number of people to define urban (1000, 1500).}

\item{pop_raster}{A raster::raster in projection}
}
\value{
A vector with c(numerator, denominator) of pixel counts.
}
\description{
This assumes that the raster values are density per square kilometer.
It counts raster values above a cutoff.
}
