% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{urban_fraction_population}
\alias{urban_fraction_population}
\title{Urban fraction from a population raster}
\usage{
urban_fraction_population(density_raster, urban_per_kilometer_sq)
}
\arguments{
\item{urban_per_kilometer_sq}{A number of people to define urban (1000, 1500).}

\item{pop_raster}{A raster::raster in projection}
}
\value{
A vector with c(numerator, denominator) of pixel counts.
}
\description{
This assumes that raster values are population in that grid location,
not density of the population in the vicinity of that grid location.
It divides population by the area of the grid pixel.
}
