% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{write_aligned_raster}
\alias{write_aligned_raster}
\title{Write a raster layer to a file.}
\usage{
write_aligned_raster(layer, layer_id, local_directory = "inst/extdata")
}
\arguments{
\item{layer}{is the RasterLayer to write.}

\item{layer_id}{is a list describing the layer with
(source = (HRSL, BIMEP, LandScan, WorldPop), grid = (HRSL, LandScan, Worldpop))
and resolution = (100, 1000).}

\item{local_directory}{is where we put the data. This data
will be written to a subdirectory of that.}
}
\description{
Our intermediate datasets go into a directory with a given format.
We write these files into an `aligned` subdirectory of the local directory.
}
