% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{summary_statistics}
\alias{summary_statistics}
\title{Compute basic summary statistics.}
\usage{
summary_statistics(population, density, urban_per_kilometer_sq)
}
\arguments{
\item{population}{raster::raster of population data.}

\item{density}{raster::raster that is density estimated population data}
}
\value{
A list with the statistics, which are maximum value, total
    population, pareto fraction, urban fraction, and NA fraction.
    Columns are `maximum`, `total`, `empty_fraction`, `pareto_fraction`,
    `urban_fraction`, `na_fraction`.
}
\description{
Computes total population, maximum population density,
percent empty, percent urban, and pareto number.
Total population makes sense here because the area is
surrounded by water, so that every filled pixel is part
of this administrative unit.
}
