% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{truth_table}
\alias{truth_table}
\title{Constructs a count of how many entries are in which truth table quadrant.}
\usage{
truth_table(condition, expected, observed)
}
\arguments{
\item{condition}{A function which, when applied to inputs, gives true and false.}

\item{expected}{The gold standard for what's true and false.}

\item{observed}{What is observed as true and false.}
}
\description{
The behavior for NA is important here. If the expected value is NA,
then this truth table ignores that entry, even if it's in the observed.
We do this because we are looking at map data where NA represents places
outside the map. Meanwhile, if the observed is NA, then we count that
as false because it won't meet the condition.
}
\seealso{
\link{accuracy_profile}
}
