% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{urban_fraction}
\alias{urban_fraction}
\title{Calculate urban fraction of a raster::raster.}
\usage{
urban_fraction(population_array, urban_cutoff = 1000)
}
\arguments{
\item{population}{raster::raster of population data}
}
\value{
Fraction, out of 1, of the land area that
    is urban.
}
\description{
The urban fraction is the fraction of land area
that has over 1000 people per square kilometer.
}
